voice_agent_outbound/pyproject.toml
[project]
name = "voice-agent-outbound"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[silero,turn-detector]==1.2.18",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "tzdata>=2025.2",
    "httpx>=0.27.2",
]
